All units implemented two operator overloads that did not really make sense. Changed this for all units. Updated tests.
Example 1:
Force operator *(Force left, Force right);

What is the result of multiplying two forces? Although it can be done, I don't see the need for it and will remove it until a need is found.

Example 2:
-        public static Force operator /(Force left, Force right)
+        public static double operator /(Force left, Force right)

The result of dividing one force with another should be the ratio, not a new force.