Fail localization test with inconclusive.
To avoid breaking the build, which happens for almost all pull requests of new units that do not have the Russian translations.
Using inconclusive flags the test so it is easy to spot, but does not prevent the test from succeeding overall.
Consider moving back to failing when a better solution for translations and/or pull request reviews is found.