Annotations: Update R# annotations to remove from binary
R#9 introduced new templates for annotations that by default excludes
them from the binary. This helps avoid conflicts with other nugets
also exporting the same annotations and the same namespace.

If we want to expose these attributes to the using code, it is
recommended to use External Annotations .xml file and add that to the
nuget package.

Fixes #62 Use official Jetbrains ReSharper annotations nuget