UnitsNet.Net35: Add pre-build step to regenerate code
Ideally this was a solution-wide pre-build step, but only way to achieve
that is to add a new project and let the others depend on it, but seems
a bit overkill. If you edit the source code (shared among the core
projects), then the Net35 project will rebuild when you build the
solution.