Add strong name signing key
Previously kept outside the repository, but according to MSDN's
recommendations for open source, this file should be made available so
others can recompile it and drop in a replacement without updating the
entire chain of dependencies.

Was able to convert the existing key .pfx file to a .snk, so that we don't
break compatibility with any signed nuget packages that have taken
dependency on the signed UnitsNet package.

See:
https://github.com/dotnet/corefx/blob/master/Documentation/project-docs/strong-name-signing.md