Remove WindowsRuntimeComponent tests
That target is not compatible with testing NUnit test framework and
we can't reference it from .NET or .NetCoreApp.
Up until now we had a WindowsRuntimeComponent.ForTests project, which targeted
.NET but had the same conditional compilation as WRC.
The problem, however, is that it can't access WRC specific attributes and types,
so that approach started to fail already.

Instead, we rely on the testing done with full framework since the differences
are mainly on public signatures and the compilation should do most of the
necessary checking for us.