Add unit converter WPF sample app
To illustrate how to enumerate quantities, units and to convert
between them without hard coding any quantities.

This sample also illustrates that there is currently need for reflection
in order to enumerate units for a selected quantity, as well as
getting the abbreviation for a unit, since the library does
not provide the means to retrieve these generically.