json: Use project references instead of nuget references
In order to catch problems due to breaking changes in UnitsNet that was not observed
due having a nuget dependency pointing to an older version of UnitsNet.
This way we always build against the latets (local checked out) UnitsNet code.

Caveat: The nuget dependency version can no longer be controlled and will always
point to the latest version (the one in the checked out code). This means pushing a new
Json nuget will also increase its dependency version if a newer one is available.
I don't currently know of a way to both get project reference AND control the nuget dependency
version, but I consider this a fair tradeoff as it is fair that taking on a newer version of Json lib will
require a newer version of UnitsNet too.