Remove net462 target from test projects
This will also speed up tests since it halves the amount
of tests to run.

This was earlier used to test Windows Runtime Component
code, but this has since been removed due to all the issues
with #if conditions and since the testing didn't seem to
add any extra safety.