json: Pass in QuantityValue instead of double
We use reflection to get the From(value, unit) method,
but this method was recently changed to take a value
of QuantityType instead of double, to use implicit
cast to avoid method overload explosion for all number
types.

Reflection, however, does not do implicit cast so
we update the deserialization code to pass in the proper
type.