Feature/less ambiguous exception (#479)
* Add test conditions for AmbiguousUnitParseException

These tests pass, confirming the current behavior.

* Return unit names from UnitSystem.Parse when throwing AmbiguousUnitParse

The new code assumes that `unitType` is an Enum. This seems safe since the first few lines of code in the `Parse` method throw an exception if no abbreviations are found for the given unit type. It would only return abbreviations if the `unitType` is a supported unit enum.

* Change single quote to double quote in AmbiguousUnitParse message

For style points. Also because of #472.