Merge generator scripts for WRC and NetFramework
Handle WRC special cases in code generator scripts instead of in generated code.
This removes all the #if sections and merges the "Common" partial code back into
each of the WRC and NetFramework files so it is in one place and easy to read.

- Arrange huge quantity code into #regions, easier to navigate
- Move code generation into functions, easier to read
- Add Types.psm1 with types for intellisense on parsed JSON*
- Require PowerShell 5.1 (for classes)

* VS and VSCode struggle with intellisense when importing types with
'using module', this is relatively new and will probably improve later.