Remove internal parse methods (#521)
* Remove redundant internal parse methods

ParseInternal and TryParseInternal methods are redundant
since the same logic is already in QuantityParser.

* Remove internal ParseUnit methods

Same as for Parse methods. DRY it up.

* QuantityParser: Parse using given abbreviations cache

Replace delegates for external unit parsing code with UnitParser.

* Add null checking and make Trim() more explicit

Trimming inline deep in the code is not easy to spot.
Make it explicit, just after null checking.
Fixes test that assumed ArgNullException.