WIP dotnet try and blazor
Try:
Add examples with UnitsNet nuget (does not work in web, only CLI)

TryWasm:
Added default project. The idea was to try to do some REPL of C# inside
it, but it may be better to just wait for `dotnet try` to get support
for nugets in the web for the sake of UnitsNet docs.

For a generic REPL, it may still be interesting to go this path.