Order the quantity json files alphabetically before running CodeGen
Directory.GetFiles doesn't guarantee the order of the file names.
In windows it seems to be retrieving them alphabetically sorted,
but that's not the case for Mac.

Have taken the liberty to rename `jsonFile` to `jsonFileName`, since
the return of Directory.GetFiles is an array of file names. The content
of the file itself is read at a later point.