✨ Allocate unique unit enum values (#1097)
Fixes #1068

Modify CodeGen to allocate unique unit enum values and store them in a JSON file.

This fixes the problem of breaking changes to the unit enum values when adding a new unit, because the enum names are sorted lexically and their values are implicitly generated.

### Summary of solution
Every time CodeGen is run, it loads a JSON file with the allocated unit enum values.
If a new unit has been added since last time, it will allocate a new enum value for that unit. 
The allocation considers the first 10 available enum values, starting from `1` and includes any gaps in the sequence. So if `LengthUnit` enum already has allocated values 2,3,4,6 then it considers the values 1,5,7,8,9,10,11,12,13,14. 

It then picks a random value from that available number sequence as the new unit's enum value. This helps avoid or simplify merge conflicts adding units to the same quantity, by not competing for the same "next" value. Since we randomly pick from the next 10 values, we will produce gaps, which is why we also try to fill those gaps in future allocations.

It then ensures there are now duplicate values, and throws if found. This will effectively break the build and describe what units are in conflict.

Next it saves the JSON file with the latest values, to be included in the PR for adding a new unit.

Finally it passes the values to `UnitTypeGenerator` to explicitly assign values when generating each unit enum type.

**Common/UnitEnumValues.g.json**
```json
{
  "Acceleration": {
    "CentimeterPerSecondSquared": 1,
    "DecimeterPerSecondSquared": 2,
    "FootPerSecondSquared": 3
  },
  "Length": {
    // ...
  }
}
```

```cs
   public enum AccelerationUnit
    {
        // Before
        Undefined = 0,
        CentimeterPerSecondSquared,
        DecimeterPerSecondSquared,
        FootPerSecondSquared,
        
        // After
        Undefined = 0,
        CentimeterPerSecondSquared = 1,
        DecimeterPerSecondSquared = 2,
        FootPerSecondSquared = 3,
```

### Changes
- Allocate new unit enum values before generating code
- Update unit type generator to assign allocated enum values
- Persist allocated unit enum values to `/Common/UnitEnumValues.g.json`
- Throw exception if duplicate values are detected.
- Pick from 10 random available values to avoid or simplify merge conflicts when merging multiple pull requests that add units to the same quantity, instead of competing for the same value.