Added BarPerSecond and BarPerMinute, with 'Milli' to PressureChangeRate. As Bar is commonly used for hydraulics and I could really use it to display pressure changes in our 'normal' unit. I don't know any russian, but i copy&pasted the abbreviations from the other units, verification needed?
Added 'Milli' to AmperePerSecond and added AmperePerMinute with 'Milli' to ElectricCurrentGradient to display some slower and minor changes in current.
Added the missing BaseUnits in AmperePerSecond and enabled SupportSIUnitSystem in the tests. I hope that is Ok?
(As it uses Ampere and Second it should as of my understanding of UnitsNet)