Add radiation exposure (#1348)
Added radiation exposure along with two units - Coulomb per kilogram
(SI-based) and Roentgen.

Coulombs per kilogram come with four prefixes (pico, nano, micro and
milli), which is two less (with kilo and mega missing) than the
definition of Coulomb itself in ElectricCharge.json. This is due to the
fact that C/kg is quite large and there are not many practical usecases
for anything beyond than that.

Roentgens come with two prefixes (milli and micro). Most often than not,
Roentgens are used 'as is', without prefixes, e.g. in case of the famous
20000R/h during Chernobyl disaster, but in some sources milli and micro
are also used (example articles below).

https://www.epa.gov/radnet/about-exposure-and-dose-rates
https://www.ncbi.nlm.nih.gov/pmc/articles/PMC7155967/

Wikipedia pages:

https://en.wikipedia.org/wiki/Radiation_exposure
https://en.wikipedia.org/wiki/Roentgen_(unit)

---------

Co-authored-by: Piotr Stenke <piotr.stenke@nexpertis.pl>