Electric
Rearranged some of the electrical units, all with the prefix "Electric", and labeled the following classes as obsolete:
* ApparentEnergy -> ElectricApparentEnergy
* ApparentPower -> ElectricApparentPower
* Capacitance -> ElectricCapacitance
* ReactiveEnergy -> ElectricReactiveEnergy
* ReactivePower -> ElectricReactivePower
Added classes for ElectricReactance and ElectricSusceptance.
Labeled ElectricPotentialAc and ElectricPotentialDc as obsolete per discussion.
Labeled ElectricAdmittance as obsolete due to it being a complex number.
Decided against adding ElectricImpedance for the same reason.
Added math to ElectricCurrent.extra.cs and ElectricPotential.extra.cs to support ElectricApparentPower
Added math for ElectricApparentPower.extra.cs
Minor correction to formula in comment of Power.extra.cs