v6-Electric
Rearranged some of the electrical units, all with the prefix "Electric", and labeled the following classes as obsolete:
* ApparentEnergy -> ElectricApparentEnergy
* ApparentPower -> ElectricApparentPower
* Capacitance -> ElectricCapacitance
* ReactiveEnergy -> ElectricReactiveEnergy
* ReactivePower -> ElectricReactivePower
Added classes for ElectricReactance and ElectricSusceptance.
Labeled ElectricPotentialAc and ElectricPotentialDc as obsolete per discussion.
Labeled ElectricAdmittance as obsolete due to it being a complex number.
Revised ObsoleteText messages to be more useful.
Added correct testing code.
Added ElectricImpedance class; even though it's a complex number, it "rounds out" the other five units of:
* Admittance
* Conductance
* Reactance
* Resistance
* Susceptance