Add a few more benchmarks (#1449)
In order to better compare the performance between the existing v6 and
the one with the [Fractions](https://github.com/danm-de/Fractions) I've
been using these benchmarks. They are not _very polished_ (and I haven't
had the time for the string-related stuff) but they do the trick.

You don't have to merge them if you don't want to, but at least they are
here (so I can reference the code).

I'll probably add the results from my local runs in a comment (or more
like a bunch of comments) later on (my previous results got wiped by the
build script 😆) - but I was hoping that we could update the packages and
add net9.0 to the list of targets first (there are [lots of a cool
stuff](https://github.com/danm-de/Fractions/pull/90) in net9).

PS I'm not sure if that old `gh-action` would still work or not, but
even adding the results by hand
([here?](https://github.com/angularsen/UnitsNet/tree/gh-pages/benchmarks/))
is an option.

Here's what I did for the
[Fractions](https://github.com/danm-de/Fractions/blob/master/benchmarks/Readme.md)
(the `Readme.md` was 90% AI generated).