Added fluid resistance quantity (#1474)
https://en.wikipedia.org/wiki/Vascular_resistance

1. I know that negative powers are not welcome, but for dyn·s/cm⁵ and
dyn·s·cm⁻⁵ units I added 2 options because both are equally common.
2. The unit MPa·s/m³ comes with the mega- prefix because that is the
only combination that is used. I have not seen it as just Pa·s/m³ or in
any other form.
3. WhodUnit actually has 3 different abbreviations, but they all mean
the same thing. For mmHg·min/l, translating popular units into Russian
was not difficult, but for WU and HRU I had to improvise. WU - Wood
Units, which in turn is translated as "Единицы Вуда" (Units of Wood), so
two words in the Russian version have changed places, and the capital
letters form ЕВ (WU = ЕВ). For HRU I translated Hybrid Resistance Unit
and used the capital letters of the Russian version. Translated as
"Единицы Гибридного Сопротивления" (HRU = ЕГС)

---------

Co-authored-by: Alexey Yagelo <Alexey.Yagelo@medi.report>